requirements.txt
discord.py
apscheduler
python-dotenv
google-genai
requests
pdfplumber
